calculo

+++ boletin2_2/src/boletin2_2/Boletin2_2.java
@@ -9,7 +9,7 @@ public class Boletin2_2 {
     
     public static void main(String[] args) {
        Scanner ler= new Scanner (System.in);
-       double graosC, graosF, graosK;
+       float graosC, graosF, graosK;
        System.out.println("Introduce graos centígrados");
        graosC=ler.nextDouble();
        graosF= 32+(9*graosC/5);
